// src/frontend/src/declarations/backend.did.d.ts
export type BodyType = { 'suv' : null } |
  { 'van' : null } |
  { 'convertible' : null } |
  { 'coupe' : null } |
  { 'sedan' : null } |
  { 'pickup' : null } |
  { 'wagon' : null } |
  { 'hatchback' : null };
export interface Cell { 'value' : Value, 'name' : string }
export type Condition = { 'new' : null } |
  { 'certifiedPreOwned' : null } |
  { 'used' : null };
export type Drivetrain = { 'awd' : null } |
  { 'fwd' : null } |
  { 'rwd' : null } |
  { 'fourWd' : null };
export type Error = { 'FrontendOriginsNotConfigured' : null } |
  {
    'MixedSsoSources' : {
      'otherKeys' : Array<string>,
      'ssoKeys' : Array<string>,
    }
  } |
  { 'Stale' : { 'ageNs' : bigint } } |
  { 'MalformedCandid' : null } |
  { 'AmbiguousAttribute' : { 'field' : string, 'sources' : Array<string> } } |
  { 'NoAttributes' : null } |
  { 'UnknownNonce' : null } |
  { 'UntrustedSsoSource' : { 'domain' : string } } |
  { 'MissingField' : string } |
  { 'FrontendOriginMismatch' : { 'got' : string, 'expected' : Array<string> } };
export type ExternalBlob = Uint8Array;
export type FuelType = { 'petrol' : null } |
  { 'hybrid' : null } |
  { 'diesel' : null } |
  { 'plugInHybrid' : null } |
  { 'electric' : null };
export interface Result { 'hasMore' : boolean, 'rows' : Array<Array<Cell>> }
export type Result__1 = { 'ok' : null } |
  { 'err' : Error };
export type SortField = { 'mileage' : null } |
  { 'year' : null } |
  { 'price' : null };
export type SortOrder = { 'asc' : null } |
  { 'desc' : null };
export type Transmission = { 'automatic' : null } |
  { 'manual' : null } |
  { 'semiAutomatic' : null };
export type UserRole = { 'admin' : null } |
  { 'user' : null } |
  { 'guest' : null };
export type Value = { 'int' : bigint } |
  { 'nat' : bigint } |
  { 'float' : number } |
  { 'bool' : boolean } |
  { 'null' : null } |
  { 'text' : string };
export interface Vehicle {
  'id' : VehicleId,
  'vin' : string,
  'model' : string,
  'sellerNotes' : string,
  'title' : string,
  'featured' : boolean,
  'mileage' : bigint,
  'drivetrain' : Drivetrain,
  'make' : string,
  'color' : string,
  'year' : bigint,
  'sellerContact' : string,
  'description' : string,
  'transmission' : Transmission,
  'fuelType' : FuelType,
  'price' : bigint,
  'bodyType' : BodyType,
  'location' : string,
  'photos' : Array<ExternalBlob>,
  'condition' : Condition,
}
export interface VehicleFilter {
  'make' : [] | [string],
  'year' : [] | [bigint],
  'maxPrice' : [] | [bigint],
  'fuelType' : [] | [FuelType],
  'minPrice' : [] | [bigint],
  'bodyType' : [] | [BodyType],
  'condition' : [] | [Condition],
}
export type VehicleId = bigint;
export interface VehicleQuery {
  'sortField' : SortField,
  'sortOrder' : SortOrder,
  'filter' : VehicleFilter,
}
export interface _ImmutableObjectStorageCreateCertificateResult {
  'method' : string,
  'blob_hash' : string,
}
export interface _ImmutableObjectStorageRefillInformation {
  'proposed_top_up_amount' : [] | [bigint],
}
export interface _ImmutableObjectStorageRefillResult {
  'success' : [] | [boolean],
  'topped_up_amount' : [] | [bigint],
}
export interface _SERVICE {
  '_immutableObjectStorageBlobsAreLive' : ActorMethod<
    [Array<Uint8Array>],
    Array<boolean>
  >,
  '_immutableObjectStorageBlobsToDelete' : ActorMethod<[], Array<Uint8Array>>,
  '_immutableObjectStorageConfirmBlobDeletion' : ActorMethod<
    [Array<Uint8Array>],
    undefined
  >,
  '_immutableObjectStorageCreateCertificate' : ActorMethod<
    [string],
    _ImmutableObjectStorageCreateCertificateResult
  >,
  '_immutableObjectStorageRefillCashier' : ActorMethod<
    [[] | [_ImmutableObjectStorageRefillInformation]],
    _ImmutableObjectStorageRefillResult
  >,
  '_immutableObjectStorageUpdateGatewayPrincipals' : ActorMethod<[], undefined>,
  '_initialize_access_control' : ActorMethod<[], undefined>,
  '_internet_identity_sign_in_finish' : ActorMethod<[], Result__1>,
  '_internet_identity_sign_in_start' : ActorMethod<[], Uint8Array>,
  'assignCallerUserRole' : ActorMethod<[Principal, UserRole], undefined>,
  'createVehicle' : ActorMethod<[Vehicle], VehicleId>,
  'execute' : ActorMethod<[string], Result>,
  'getApiDoc' : ActorMethod<[], string>,
  'getCallerUserRole' : ActorMethod<[], UserRole>,
  'getVehicle' : ActorMethod<[VehicleId], [] | [Vehicle]>,
  'isCallerAdmin' : ActorMethod<[], boolean>,
  'listVehicles' : ActorMethod<[], Array<Vehicle>>,
  'schema' : ActorMethod<[], string>,
  'searchVehicles' : ActorMethod<[VehicleQuery], Array<Vehicle>>,
}
export declare const idlService: IDL.ServiceClass;
export declare const idlInitArgs: IDL.Type[];
export declare const idlFactory: IDL.InterfaceFactory;